% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightswin.R
\name{determinconstws}
\alias{determinconstws}
\title{determinconstws}
\usage{
determinconstws(pii, x, bi, maxit = 10000)
}
\arguments{
\item{pii}{First order inclusion probabilities}

\item{x}{Variable(s) of interest}

\item{bi}{Conditional bias}

\item{maxit}{Maximum number of iterations for the research of the minimum}
}
\value{
Computes the winsorisation constant associated to the standard winsorized estimator
}
\description{
determinconstws
}
