% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTcondbiasest.R
\name{HTcondbiasest}
\alias{HTcondbiasest}
\title{HTcondbiasest
The function to estimate the conditional biases for the HT estimator of a total, for the non-stratiﬁed sampling designs. For one or several given variables of interest as input, this function returns a vector with the estimates of the conditional bias of the HT estimator where each row corresponds to a sample unit.}
\usage{
HTcondbiasest(
  data,
  varname = NULL,
  gn,
  method = c("si", "poisson", "rejective"),
  pii = NULL,
  di = NULL,
  id = NULL,
  remerge = T
)
}
\arguments{
\item{data}{Original data set with the sample, the variable(s) of interest}

\item{varname}{Name(s) of the variable(s) of interest}

\item{gn}{Population size}

\item{method}{Sampling design : si for simple random sampling, poisson, rejective}

\item{pii}{First order inclusion probabilities}

\item{di}{Inverse of the first order inclusion probabilities}

\item{id}{Primary key to keep as an identifier of every unit}

\item{remerge}{True/False to remerge the resulting conditional biases with the original data set}
}
\value{
Returns a data set with the conditional bias of each sampled unit
}
\description{
HTcondbiasest
The function to estimate the conditional biases for the HT estimator of a total, for the non-stratiﬁed sampling designs. For one or several given variables of interest as input, this function returns a vector with the estimates of the conditional bias of the HT estimator where each row corresponds to a sample unit.
}
