\name{jeu_donnees}
\alias{jeu_donnees}
\title{Prices and surface of 537 274 buildings in France}



\description{
 A dataset containing the price and the surface of 537 274 flats, houses or buildings. This data set is an extraction of the open data set providing by the French Tax Service.
The database does not contain personal data, such as the name of the seller or the buyer of a property. It only contains information about transactions: type of property sold, area, sale price and so on.

 The variables are as follows:
 \itemize{
   \item{Commune :}{town of the building}
   \item{Valeur.fonciere_num : } {the price of the builing in euros}
   \item{Surface.reelle.bati :} {surface of the building}
   \item{Type.local_c :} {type of building : house,flat,...}
   \item{strata : } {numeric strata corresponding to the Type.local_c}
}

The original data set is available at \url{https://www.data.gouv.fr/en/datasets/demandes-de-valeurs-foncieres/}
}
