% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustest.R
\name{robustest}
\alias{robustest}
\title{robustest}
\usage{
robustest(
  data,
  varname = NULL,
  gn,
  method = c("si", "poisson", "rejective"),
  pii
)
}
\arguments{
\item{data}{Original data set with the sample, the variable(s) of interest}

\item{varname}{Name(s) of the variable(s) of interest}

\item{gn}{Population size}

\item{method}{Sampling design : si for simple random sampling, poisson, rejective}

\item{pii}{First order inclusion probabilities}
}
\value{
Computes the robust estimator of Beaumont et al.(2013) using the conditional bias and the minmax criterion to compute the tuning constant
}
\description{
robustest
}
