% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightswin.R
\name{robustweights.r}
\alias{robustweights.r}
\title{robustweights}
\usage{
robustweights.r(
  data,
  varname = NULL,
  gn,
  method = "si",
  pii,
  typewin = NULL,
  maxit = 10000,
  remerge = T
)
}
\arguments{
\item{data}{Original data set with the sample, the variable(s) of interest}

\item{varname}{Name(s) of the variable(s) of interest}

\item{gn}{Population size}

\item{method}{Sampling design : si for simple random sampling, poisson, rejective}

\item{pii}{First order inclusion probabilities}

\item{typewin}{Winsorized estimator : Beaumont et al., Standard or Dalen-Tambay}

\item{maxit}{Maximum number of iterations for the research of the minimum}

\item{remerge}{True/False to remerge the conditional bias with the original data set}
}
\value{
Computes the robust weights associated to the winsorized estimator
}
\description{
robustweights
}
