% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_HTcondbiasest.R
\name{strata_HTcondbiasest}
\alias{strata_HTcondbiasest}
\title{strata_HTcondbiasest
The function to estimate the conditional biases for the HT estimator of a total, for the stratified sampling designs. For one or several given variables of interest as input, this function returns a vector with the estimates of the conditional bias of the HT estimator where each row corresponds to a sample unit.}
\usage{
strata_HTcondbiasest(
  data,
  strataname = NULL,
  varname = NULL,
  gnh,
  method = c("si", "poisson", "rejective"),
  pii = NULL,
  di = NULL,
  id = NULL,
  remerge = T
)
}
\arguments{
\item{data}{Original data set with the sample, the variable(s) of interest}

\item{strataname}{Name of the variable to use for stratification}

\item{varname}{Name(s) of the variable(s) of interest}

\item{gnh}{Population size in each stratum}

\item{method}{Sampling design : si for simple random sampling, poisson, rejective}

\item{pii}{First order inclusion probabilities}

\item{di}{Inverse of the first order inclusion probabilities}

\item{id}{Primary key to keep as an identifier of every unit}

\item{remerge}{True/False to remerge the resulting conditional biases with the original data set}
}
\value{
Returns a dataframe with the conditional bias of each sampled unit
}
\description{
strata_HTcondbiasest
The function to estimate the conditional biases for the HT estimator of a total, for the stratified sampling designs. For one or several given variables of interest as input, this function returns a vector with the estimates of the conditional bias of the HT estimator where each row corresponds to a sample unit.
}
