% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strata_robustest.R
\name{strata_robustest}
\alias{strata_robustest}
\title{strata_robustest}
\usage{
strata_robustest(
  data,
  strataname = NULL,
  varname = NULL,
  gnh,
  method = c("si", "poisson", "rejective"),
  pii
)
}
\arguments{
\item{data}{Original data set with the sample, the variable(s) of interest}

\item{strataname}{Name of the variable to use for stratification}

\item{varname}{Name(s) of the variable(s) of interest}

\item{gnh}{Population size in each stratum}

\item{method}{Sampling design : si for simple random sampling, poisson, rejective}

\item{pii}{First order inclusion probabilities}
}
\value{
Computes the robust estimator of Beaumont et al.(2013) by stratum using the conditional bias and the minmax criterion to compute the tuning constant
}
\description{
strata_robustest
}
