% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightswin.R
\name{strata_robustweights.r}
\alias{strata_robustweights.r}
\title{strata_robustweights}
\usage{
strata_robustweights.r(
  data,
  strataname = NULL,
  varname = NULL,
  gnh,
  method = c("si", "poisson", "rejective"),
  pii,
  typewin = NULL,
  maxit = 10000,
  remerge = T
)
}
\arguments{
\item{data}{Original data set with the sample, the variable(s) of interest}

\item{strataname}{Name of the variable to use for stratification}

\item{varname}{Name(s) of the variable(s) of interest}

\item{gnh}{Population size in each stratum}

\item{method}{Sampling design : si for simple random sampling, poisson, rejective}

\item{pii}{First order inclusion probabilities}

\item{typewin}{Winsorized estimator : Beaumont et al., Standard or Dalen-Tambay}

\item{maxit}{Maximum number of iterations for the research of the minimum}

\item{remerge}{True/False to remerge the conditional bias with the original data set}
}
\value{
Computes the robust weights associated to the winsorized estimator
}
\description{
strata_robustweights
}
