% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightswin.R
\name{tuningconst}
\alias{tuningconst}
\title{tuningconst}
\usage{
tuningconst(bi)
}
\arguments{
\item{bi}{Conditional biases of the variable of interest}
}
\value{
Computes the tuning constant associated to the Beaumont et al (2013) estimator
}
\description{
tuningconst
}
