% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightswin.R
\name{wrapper}
\alias{wrapper}
\title{wrapper
The main wrapper function}
\usage{
wrapper(
  data,
  varname = NULL,
  strataname = NULL,
  gn,
  gnh = NULL,
  est_type = c("BHR", "standard", "DT"),
  method = c("si", "poisson", "rejective"),
  pii = NULL,
  di = NULL,
  id = NULL
)
}
\arguments{
\item{data}{Original data set with the sample, the variable(s) of interest}

\item{varname}{Name(s) of the variable(s) of interest}

\item{strataname}{Name of the variable to use for stratification}

\item{gn}{Population size}

\item{gnh}{Population size in each stratum}

\item{est_type}{Type of estimator}

\item{method}{Sampling design: si for simple random sampling, poisson, rejective}

\item{pii}{First order inclusion probabilities}

\item{di}{sampling weights}

\item{id}{unit identifier}
}
\value{
Returns two dataframes: the first one gives the robust estimator and acts as a summary of the winsorisation. The second one details the weight changes of every unit.
}
\description{
wrapper
The main wrapper function
}
